+++ src/Scriptbook/Scriptbook.psm1
@@ -224,6 +224,8 @@ if ($depends)
             continue
         }
 
+        $skipPublisherCheck = if ($dependency.ContainsKey('SkipPublisherCheck')) { $dependency.SkipPublisherCheck } else { $false }
+
         $minimumVersion = if ($dependency.ContainsKey('MinimumVersion')) { $dependency.MinimumVersion } else { '' }
         $maximumVersion = if ($dependency.ContainsKey('MaximumVersion')) { $dependency.MaximumVersion } else { '' }
 
@@ -302,7 +304,7 @@ if ($depends)
                 {
                     Write-Warning "Module $($dependency.Module) not installed by Install-Module, cannot update module via Update-Module, using forced Install-Module"
                     Write-Verbose "Installing module $($dependency.Module)"
-                    Install-Module -Name $dependency.Module -Force -Repository $repository -Scope CurrentUser -MinimumVersion $minimumVersion -MaximumVersion $maximumVersion -AllowClobber @extraParams
+                    Install-Module -Name $dependency.Module -Force -Repository $repository -Scope CurrentUser -MinimumVersion $minimumVersion -MaximumVersion $maximumVersion -AllowClobber -SkipPublisherCheck:$skipPublisherCheck @extraParams
                 }
             }
         }
@@ -310,7 +312,7 @@ if ($depends)
         {
             Write-Verbose "Installing module $($dependency.Module)"
             # TODO !!EH using -Force to install from untrusty repositories or do we need to handle this via Force attribute
-            Install-Module -Name $dependency.Module -Force -Repository $repository -Scope CurrentUser -MinimumVersion $minimumVersion -MaximumVersion $maximumVersion -AllowClobber @extraParams
+            Install-Module -Name $dependency.Module -Force -Repository $repository -Scope CurrentUser -MinimumVersion $minimumVersion -MaximumVersion $maximumVersion -AllowClobber -SkipPublisherCheck:$skipPublisherCheck @extraParams
         }
 
         if ($dependency.ContainsKey('Args'))