+++ docs/release-notes.md
@@ -4,7 +4,26 @@
 ![PowerShell Gallery](https://img.shields.io/powershellgallery/dt/Scriptbook.svg?label=PSGallery%20Downloads&logo=PowerShell&style=flat-square)
 [![Build Status](https://dev.azure.com/tedon/TD.Deploy/_apis/build/status/ehagen.Scriptbook?branchName=master)](https://dev.azure.com/tedon/TD.Deploy/_build/latest?definitionId=52&branchName=master)
 
-## [0.5.*]
+## [0.5.01]
+
+```plain
+### Added
+
+- Added Action Always support
+- Added Start-Workflow ErrorAction variable support
+- Added Get-ActionState
+- Added skipPublisherCheck support to Import-Module
+- Added Get-EnvironmentVar
+- Added Test-PSProperty
+- Added Get-PSProperty
+
+### Changed
+
+- Cleanup Workflow Report dependencies
+
+```
+
+## [0.5.0]
 
 ```plain
 ### Added